// frontend/client/src/app/team.ts
export interface Team {
  id: number;
  name: string;
  conference: string;
  division: string;
  wins: number;
  losses: number;
  confrank: number;
  pts: number;
  asts: number;
  reb: number;
  opppts: number;
  ptsrank: number;
  astsrank: number;
  rebrank: number;
  oppptsrank: number;
  color: string;
  secondcolor: string;
}